git/stats: Get rid of spawning git-count-objects(1)
We have now converted almost all of the information and more about
objects to be derived via in-process logic instead of relying on the
external git-count-objects(1) command. There are only two pieces of
information left, both of which aren't used anywhere:

    - Prunable objects are objects that exist both as a loose and as a
      packed object. Theoretically speaking we shouldn't ever generate
      such objects in the first place as we also `git repack -d`, and
      this command will remove objects we're putting into a packfile.

    - Packed objects is the number of objects that exist in total. This
      is a bit unfortunate to go away, but we still expose the total
      packfile size.

While the second information could be useful, it doesn't feel useful
enough to argue for a whole invocation of git-count-objects(1).

Stop executing git-count-objects(1) and remove those two bits of info.
This should make it a whole less taxing to derive this information from
the repository. Ultimately, this will allow us to just use `ObjectsInfo`
for our housekeeping logic to feed into the heuristical optimization
strategy.