Support health checking connections locally only
HealthManager is responsible for health checking connections to
Gitaly in Praefect. Doing so, it also persists the information in
the database so the other Praefect's can take the connection status
of other Praefects in consideration when promoting primaries and so
forth. As we are about to add a second set of connections behind a
feature flag, we want to health check the connections so we don't
route traffic to them if they are unhealthy but we don't want to the
status of the feature flagged connections affect the health consensus
used for the elections. This commit makes this possible with the health
manager. If the HealthManager is not provided a database handle, it will
simply locally check the connections without persisting any information
in the database.