operations: Fix flaky UserRebaseConfirmable test caused by timeout
One of our tests for UserRebaseConfirmable is using a timeout to wait
for the second response to be received. There is no reason why that
happens though as we're not testing for anything related to timeouts.
Furhtermore, this timeout of 1 second causes tests on macOS to be flaky.

Remove the timeout to make the test more stable.