structerr: Allow adding error details
While the structured errors are a great addition to pass structured
information about why a certain RPC has failed to the caller, they are
quite verbose to use on our side. Add a helper function in the structerr
package that allows the user to append error details to an error to
remove most of the burden.