cleanup: Fix flaky test for invalid requests
One of our tests that verifies behaviour when sending an invalid
repository as input is flaky. This is because the `doStreamingRequest()`
helper function always sends multiple requests to the server, which
means that if the server already rejects the first message it can be
that it will cancel the stream before we have succeeded to send the
second request.

Fix this by refactoring `doStreamingRequest()` to receive an arbitrary
number of requests as input. Like this, the tests for invalid requests
can only send a single request that will fail validation, while other
tests can continue to stress our ability to reassemble the object map.