objectpool: Merge functions to implement creation of pools
Creating pools really is quite straight forward and not all that
complex. Merge the functionality into a single function so that we can
be reasonably sure that all code paths will always perform a full
creation of object pools instead of only calling one of the previously
available subfunctions.

Note that this also changes the comment to say that the source
repository will _not_ join the object pool, whereas it previously stated
the opposite. The comment had been wrong though: the sorce repository
does indeed not join.