objectpool: Refactor setup test helper to return localrepo
We always need a `*localrepo.Repo` when working with object pools, but
the setup helper returns a `*gitalypb.Repository`. This means that all
tests have to manually construct it.

Refactor the helper to construct and return a `*localrepo.Repo`. While
at it, drop the call to `t.Cleanup()` to remove the pool again as it
will be deleted anyway when the temporary directory is pruned.