Remove gitaly_always_log_feature_flags flag
In a prior version, we introduced feature flag logging as a part of gRPC
logging middleware. That was a retry attempt, because we blew up our
centralized logging system due to huge volume of feature flag logs. As
a result, we attach feature_flag files to failed RPCs only. We don't
include them in successful RPCs unless gitaly_always_log_feature_flags
flag is on.

After turned ou the flag on production, there was no impact on our
logging system. This commit is to remove the aforementioned flag.
Afterward, this field is always attached.