git: Make function to get global Git configuration public
Most of our tests for the `git` package are in a separate `git_test`
package due to cyclic dependencies. We thus cannot access the private
function `git.ExecCommandFactory.globalConfiguration()`, which makes it
hard for use to test the way we assemble the Git configuration.

Expose the function to be publicly visible. While at it, change its
return type to be an array of `ConfigPair`s instead of the interface to
adher to the rule "accept interface, return concrete type". While this
makes things a bit awkward to use when we want to append the config to
arrays of the interface type, it is ultimately easier to test.