config: Allow empty values for Git configuration
The validity checks for our Git configuration is too strict as it does
not allow for empty values. This breaks some valid usecases, like for
example when an administrator wants to unset specific keys.

Drop the check for empty values and accept such keys. We have already
asserted in the preceding commit that Git handles such Git configuration
just fine.

Changelog: fixed