remote: Improve performance of UpdateRemoteMirror test setup
The test setup of our UpdateRemoteMirror is really slow depending on the
machine type. In CI we frequently see that this test thus needs multiple
minutes alone to execute.

The root cause of this is that we call more than a thousand Git commands
to set up the test data because we call `gittest.WriteCommit()` about
250 times. And each of these calls will call git-hash-object(1) to write
a blob, git-write-tree(1) to write the tree, git-commit-tree(1) to write
the commit, and then ultimately git-update-ref(1) to update the target
reference. So we execute more than a 1000 Git commands, likely more than
that.

Now doing this is kind of a requirement to test divergent refs, where we
have a cutoff point at 100 divergent references. But most of the tests
actually write the same commit over and over again, which is of course
not all that efficient. Furthermore, instead of updating references via
separate calls to git-update-ref(1), we can just use the `updateref`
package to instead do so with a single Git command.

While this complicates the test setup, it greatly improves performance
four-fold on my machine from about 2.8 seconds to 0.7 seconds. On weaker
systems like CI machines or on systems where running processes is slower
than on Linux, like macOS, the effect should be a whole lot more
pronounced.