backup: Do not use seeded repository for tests
The backup tests are currently one of the most expensive tests we have
in our test suite. This is because we create quite a few copies of the
seed repository via git-archive(1), and that cost quickly adds up.

We don't really depend on the repository's contents, we just require
that it has _some_ contents. So let's instead seed the repository with
custom data at runtime. This improves the time to execute those tests
from about 5.6 seconds to 1.4 seconds on my machine.