housekeeping: Embed repository info in heuristical optimization strategy
We're manually assembling all kinds of information on a repository by
calling the respective functions in the `git/stats` package and then
extract the relevant bits of information. Now that we have a single
unified structure `RepositoryInfo` though that is not necessary.

Refactor the code to embed a `RepositoryInfo` directly.