housekeeping: Always log repository info in `OptimzeRepository`
We're logging repository information in the `RepackRepository` function.
This means that this information is logged both in the `RepackFull` and
`RepackIncremental` RPCs, but we only log it via `OptimizeRepository`
when we decide to actually repack objects. Furthermore, we essentially
derive the repository information twice in `OptimizeRepository`: once to
inform the heuristical optimization strategy, and once to log it in case
we decide to repack objects.

Fix this by logging the information when creating the heuristical
optimization strategy. This ensures we always log it regardless of
whether we repack objects or not, and it means we don't have to derive
the information twice. To retain previous semantics for `RepackFull` and
`RepackIncremental` we now explicitly call `LogRepositoryInfo()` there.

Changelog: fixed