objectpool: Merge FromProto and NewObjectPool
The lifetime of object pools is not quite clear right now due to the
number of creators we have. Reduce the number a bit by merging the
`FromProto()` and `NewObjectPool()` constructors. While this is a bit
more annoying in the context of tests, a Protobuf representation is what
we'd typically see in the RPCs.