objectpool: Refactor `Create()` to act as a constructor
To create an object pool you first have to create the `ObjectPool`
object and then call `Create()` on it. This is kind of weird and
restricts us from putting additional sanity checks on both of the other
constructors `FromRepo()` and `FromProto()` to check whether the object
pool exists in the first place.

Refactor the code so that `Create()` is a constructor that returns an
`ObjectPool` object.