objectpool: Create object pool in test helper
The objectpool testhelelper `setupObjectPool()` only creates the object
pool object, but doesn't actually create the object pool itself. This
makes the test setup more complex than required as most tests want to
create the object pool anwyay. And those where we don't want to do this
we can just call `objectPool.Remove()` and be done.

Refactor the test helper to create the pool repository to make the test
setup less repetitive in general.