objectpool: Drop unused `Init()` function
The `Init()` function is used to create a new empty object pool. But
object pools are in fact never created as new, empty repositories but
are always created from an initial object pool repository. As a
consequence, this function is now unused.

Remove the unused code to further clarify the lifetime of object pools.