repository: Create packfiles more efficiently in MIDX tests
The `addPackFiles()` helper function generates a configurable number of
packfiles into the target repository. In order to arrive at different
packfile sizes, the function will create `n` commits for the `n`th
packfile. This means that we in total execute `4n(n+1)/2` many Git
commands just to write commits. In total this means we execute over a
1000 Git commands in `TestMidxRepackExpire()`.

As the intent was simply to generate packfiles of different sizes, let's
do this by just writing a single commit per packfile that has a commit
message length that grows. This significantly cuts the runtime of the
test by almost half. Before this change:

--- PASS: TestMidxRepackExpire (3.83s)
    --- PASS: TestMidxRepackExpire/Test_repack_expire_with_3_added_packs (0.11s)
    --- PASS: TestMidxRepackExpire/Test_repack_expire_with_5_added_packs (0.22s)
    --- PASS: TestMidxRepackExpire/Test_repack_expire_with_11_added_packs (0.61s)
    --- PASS: TestMidxRepackExpire/Test_repack_expire_with_20_added_packs (1.71s)

After this change:

--- PASS: TestMidxRepackExpire (2.09s)
    --- PASS: TestMidxRepackExpire/Test_repack_expire_with_3_added_packs (0.08s)
    --- PASS: TestMidxRepackExpire/Test_repack_expire_with_5_added_packs (0.12s)
    --- PASS: TestMidxRepackExpire/Test_repack_expire_with_11_added_packs (0.27s)
    --- PASS: TestMidxRepackExpire/Test_repack_expire_with_20_added_packs (0.49s)

On other systems where spawning Git commands is more expensive than on
Linux the effect is going to be more pronounced.