git/stats: Speed up PerformHTTPPush tests
One of the subtests for `PerformHTTPPush` is creating a thousand
commits. Notably though, each of these commits is exactly the same, so
we essentially execute 4000 Git commands just to recompute the same
results.

Change the logic to just compute the commit once. This significantly
speeds up the time to execute these tests. Before:

--- PASS: TestPerformHTTPPush (3.22s)
    --- PASS: TestPerformHTTPPush/single_revision (0.02s)
    --- PASS: TestPerformHTTPPush/many_revisions (3.18s)
    --- PASS: TestPerformHTTPPush/branch_deletion (0.01s)
    --- PASS: TestPerformHTTPPush/failing_delete (0.01s)

After:

--- PASS: TestPerformHTTPPush (0.08s)
    --- PASS: TestPerformHTTPPush/single_revision (0.01s)
    --- PASS: TestPerformHTTPPush/many_revisions (0.05s)
    --- PASS: TestPerformHTTPPush/branch_deletion (0.01s)
    --- PASS: TestPerformHTTPPush/failing_delete (0.01s)