operations: Refactor how we compute trees in UserApplyPatch tests
The way we compute both the base tree as well as the patched trees in
our UserApplyPatch tests is quite confusing. Refactor it to instead use
the gittest package to hopefully make the setup easier to understand.

While I was hoping for a speedup due to this refactoring there wasn't
really any noticeable effect.