structerr: Retain error codes from already-wrapped gRPC errors
After we have formatted the error, we try to check whether there is any
error in the chain of wrapped errors that already has a gRPC code so
that we can retain it. We do this by checking for `structerr.Error`s in
the error chain, which works fine in many cases as we also convert any
incoming `status.Error`s to `structerr.Error`s. But it falls short in
the case where there is for example an error created with `fmt.Errorf()`
that wraps a `status.Status`, and we lose the error code in that case.

Fix this bug by instead checking for any error in the error chain that
has a `GRPCStatus()` function.

Changelog: fixed