structerr: Allow to add multiple error details
The `structerr` package currently only allows to attach a single error
detail to each `structerr`. This doesn't make any sense though: it is
perfectly fine to attach multiple details to a gRPC error, and we
already do that with nested `structerr` errors.

Allow each `structerr` to carry an array of details to fix this design
flaw.