Parse non-existent object errors from update-ref
update-ref verifies that an object exists before pointing a
reference to it. Updater currently does not parse the error. This
commit parsers the errors in to a typed one so it can be easier
handled in the calling code.

TestUpdater_capturesStderr was updated to fail the process by writing
invalid input as the error it was relying on before is now parsed and
thus not returned in the stderr.