global: Stop wrapping potenitally-nil errors
The error-wrapping helper `ErrInternal()` et al. would return `nil` in
case the input error was `nil`. We rely on this fact at two callsites,
but this will stop working as soon as we convert all of them to instead
use the error-formatting variant `ErrInternalf()`.

Convert those callsites to explicitly handle errors.