helper: Propagate error details in error-formatting functions
When wrapping an error with the error-formatting functions that contains
error details then we don't embed those details into the new gRPC status
object. So ultimately we end up swallowing the details and will never
propagate them to the caller, which is a footgun waiting to go off. This
is doubly confusing because the non-formatting error helpers correctly
propagate the error details.

Fix this by properly embedding the error details into the generated gRPC
status, if any.