git/objectpool: Convert usage of `FullPath()` to `Path()`
Object pools provide two ways to calculate the path of a pool:
`FullPath()` returns the path without verifying whether the pool exists,
while `Path()` verifies that the target exists. The latter is what we
already use in the context of the `localrepo.Repo` package, and it gives
us more safeguards to ensure that the object pool really exists when it
ought to exist.

Refactor the code to use `Path()` instead of `FullPath()`.