ci: Add a prefix to distinguish caches by build images
As we only really create CI caches on the Debian-based tests we haven't
yet run into any issues when trying to reuse a cache generated for one
of the images for a CI job that uses a different image. We're about to
start caching results for the FIPS runner and for macOS though, so this
will become more of a problem.

Introduce a new cache prefix variable that by default refers to Debian
based images. Other jobs can easily override this prefix as required.