ci: Create Go module cache as read-writeable
By default, Go creates the module cache as read-only. While this works
alright in our CI right now, it breaks caching on macOS. This is because
the created cache would have the same read-only permissions, and macOS
is not able to extract the cache when the directories it contains cannot
be written to.

Ask Go to create the cache as read-writeable files to fix this issue.