git/objectpool: Refuse creating pools when empty target dir exists
The git-clone(1) command will happily clone into a target directory, but
only in the case it is an empty directory. This behaviour is something
we have carried over when creating object pools. It is arguably not the
right thing to do though: when we're asked to create something that
exists already in some kind of form then we really ought to refuse the
operation and return an error code.

Tighten our APIs to always refuse creating of object pools when the
target directory exists.

Changelog: fixed