repository: Refactor RepositorySize tests to only use ObjectPool service
The tests for RepositorySize in combination with object pools construct
the `objectpool.Pool` object before the pool has been created. While
this works right now, we're about to change behaviour of this so that we
require the object pool to exist when instantiating the object.

Refactor the test to exclusively use the ObjectPool services. We don't
need access to the `objectpool.ObjectPool` object in the first place,
but really only used to round-trip it to its Protobuf representation.