objectpool: Refactor Link tests to thoroughly assert errors
The tests for `LinkRepositoryToObjectPool()` only check for error codes
returned by the gRPC service instead of fully verifying the expected
error, if verifying the returned errors at all. This means that there is
quite a lot of wiggle room for what the exact returned error is.

Refactor the tests to thoroughly assert errors.