git/objectpool: Refactor pool stat logging to allow easier testing
When updating object pools via `FetchFromOrigin()` we log information
about the object pool before and after performing the updates. This is
done via `logStats()`, which will always write the log entry into the
logger set in the current context. This makes it comparatively hard to
test the function standalone.

Refactor the API to accept a `logrus.Entry` as input and add a small set
of tests for this function.