git/objectpool: Use structured logging for normal and dangling refs
When fetching into object pools we log information about normal and
dangling references in the object pool. When consuming these log
messages in something like Elastic though it is quite hard to quickly
see both of these stats. For once this is because they don't use the
same prefix. And second, we don't use structured logging and thus have
quite a bit of repetetive information for the keys.

Refactor the code to instead use structured logging. Besides being
easier to handle both in the code and as a consumer of the log entry it
should also be more efficient.