objectpool: Use `git/stats` package to derive pool info
Besides counting the number of normal and dangling references, we also
derive information on the repository's objects and references sizes. The
current caculation is implemented by spawning du(1), which is rather
ugly. Furthermore, we don't report quite a bit of other information
about object pools that might be of interest.

Refactor the code to use the `git/stats` package instead, which exposes
a whole bunch of information about repositories.