gittest: Apply global Git configuration to spawned commands
We don't apply the global Git configuration to our spawned Git commands.
As a consequence some assumptions we tend to do in our production level
code may not hold true in tests as the generated repositories didn't
have the Git configuration applied. While this isn't only true in the
context of global Git configuration but also in the per-command Git
configuration, applying the former is trivial enough.

Start applying the global Git configuration to Git commands spawned via
the `gittest` package. This fixes some issues we have when writing
commit graphs: we don't currently generate generation data in our
production code, but we do in our tests because we don't apply the
global Git configuration.

Note that this requires us to amend one of our tests which assumes that
we use replace refs by default. With the global Git configuration
applied we don't do that anymore.