git/stats: Refactor commit-graph tests to allow more involved setup
We're about to add support for parsing the generation data from commit
graphs. Part of the generation data is the overflow data that only gets
generated in repositories that have corrected committer dates that don't
fit into 31 bit. This will require a bit more involved setup than
running a single command, only, but tests for the `CommitGraphInfo`
don't allow for that right now.

Refactor the tests to use a setup function so that we can execute
arbitrarily many commands.