Makefile: Upgrade Git to v2.35.5.gl1 and v2.37.4.gl1 respectively
Upgrade Git to v2.35.5.gl1 and Git v2.37.4.gl1, respectively. This
includes bugfixes as well as fixes for CVEs that don't apply in the
context of Gitaly.

Note that the naming of the `GIT_VERSION_2_37_1` variable is now kind of
awkward as it refers to Git v2.37.4.gl1, and the same applies to the
variable `GIT_VERSION_2_35_1`. To keep changes to a minimum though we
leave these as-is.

Changelog: fixed