housekeeping: Replace commit-graph chain when missing generation data
Git will only read and write generation data from commit-graphs when the
complete chain of graphs already has generation data. As a result, when
we incrementally append a new commit-graph slice to the chain of graphs
we won't generate the commit-graph data at all.

We have essentially the same problem with bloom filters, where the
current strategy is to just rewrite the complete commit-graph chain in
case the chain is missing bloom filters. So let's do the same when they
are missing generation data.

Changelog: added