git: Simplify test to compute Ruby sidecar Git configuration
In the past we used to make the `core.fsync` setting conditional on the
Git version as it had only been introduced in Git v2.36.0. This can
still be seen by one of our tests for retrieving the Ruby sidecar Git
configuration, which still goes through some hoops to support this.

Simplify the test to not use an intercepting Git command factory as we
don't need to replace the version anymore and make the list of expected
Git configuration static.